PersistentWS: allow more control over retries
Let onClose return an object indicating whether it's OK to retry. This
lets platforms stop the reconnection loop whenever appropriate (e.g. if
the provided credentials are invalid - reconnecting doesn't help us
there).

Also add a disconnect method that tries to close the underlying
WebSocket with a certain code, but still dispatches onClose to let
upstream consumers decide whether to reconnect or not. This is useful if
some server state is tied to the lifecycle of the WebSocket itself, and
we need to simply reconnect.